+++ Biblioteca.Aplicacion/Prestamo/PrestamoCasosDeUso/RealizarPrestamoCU.cs
@@ -1,10 +1,6 @@
-//Archivo sin uso. Borrar despues de testear
-
-/*
-
 using System.Net.Cache;
-
 namespace Biblioteca.Aplicacion;
+
 
 
 public class RealizarPrestamoUseCase
@@ -55,53 +51,3 @@ public class RealizarPrestamoUseCase
 
     }
 }
-
-public class DevolverLibroUseCase
-{
-    private readonly IRepositorioPrestamo _repo;
-    public DevolverLibroUseCase(IRepositorioPrestamo repo)
-    {
-        this._repo = repo;
-    }
-    public void Ejecutar(Prestamo prestamo)
-    {
-        _repo.devolverLibro(prestamo);
-    }
-}
-
-public class ListarPrestamosUseCase
-{
-    private readonly IRepositorioPrestamo _repo;
-    public ListarPrestamosUseCase(IRepositorioPrestamo repo)
-    {
-        this._repo = repo;
-    }
-    public List<Prestamo> Ejecutar()
-    {
-        return _repo.listarPrestamos();
-    }
-}
-
-public class ListarPrestamosActivosUseCase
-{
-    private readonly IRepositorioPrestamo _repo;
-    public ListarPrestamosActivosUseCase(IRepositorioPrestamo repo)
-    {
-        this._repo = repo;
-    }
-    public List<Prestamo> Ejecutar()
-    {
-        var lista = _repo.listarPrestamos();
-        var retLista = new List<Prestamo>();
-
-        foreach (Prestamo p in lista){
-            if(p.estaDevuelto == false){
-                retLista.Add(p);
-            }
-        }
-
-        return retLista;
-    }
-}
-
-*/
